requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
h11==0.16.0
idna==3.11
influxdb-client==1.50.0
numpy==2.4.2
paho-mqtt==2.1.0
pandas==3.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reactivex==4.1.0
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
PyYAML>=6.0
PyJWT>=2.8.0
bcrypt>=4.0.0
passlib>=1.7.4
httpx>=0.27.0
pytest>=8.0.0
scikit-learn>=1.4.0
pymodbus>=3.8.0
openai>=1.63.2
google-generativeai>=0.8.0
google-genai>=1.0.0
